requirements.txt
discord.py==0.16.7
requests==2.13.0